Code Smell 19 & 20 & 21 & 22 #
19#:
Declare "port" on a separate line.

Multiple variables should not be declared on the same linejava:S1659

20#:
 Declare "host" on a separate line.

 Multiple variables should not be declared on the same linejava:S1659

21#
Remove the "port" field and declare it as a local variable in the relevant methods.

Private fields only used as local variables in methods should become local variablesjava:S1450

22#
Remove the "host" field and declare it as a local variable in the relevant methods.

Private fields only used as local variables in methods should become local variablesjava:S1450